Migrated the benchmark to a separate file
